requirements-lock-py38.txt
anyio==3.7.1
blinker==1.8.2
blis==0.7.11
catalogue==2.0.10
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
confection==0.1.5
contourpy==1.1.1
cryptography==41.0.7
cycler==0.12.1
cymem==2.0.11
distro==1.9.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl#sha256=0964370218b7e1672a30ac50d72cdc6b16f7c867496f1d60925691188f4d2510
exceptiongroup==1.3.0
Flask==2.3.3
fonttools==4.57.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.1
kiwisolver==1.4.7
langcodes==3.4.1
language_data==1.3.0
Levenshtein==0.21.1
marisa-trie==1.2.1
MarkupSafe==2.1.5
matplotlib==3.7.2
murmurhash==1.0.13
numpy==1.24.4
openai==1.98.0
packaging==25.0
pandas==2.0.3
pathlib_abc==0.1.1
pathy==0.11.0
Pillow==10.0.1
pluggy==1.5.0
preshed==3.0.10
pycparser==2.22
pydantic==1.10.22
pyparsing==3.0.9
pytest==7.4.3
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-Levenshtein==0.21.1
pytz==2025.2
rapidfuzz==3.9.7
requests==2.31.0
six==1.17.0
smart-open==6.4.0
sniffio==1.3.1
spacy==3.5.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
thinc==8.1.12
tomli==2.2.1
tqdm==4.67.1
typer==0.9.4
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
wasabi==1.1.3
Werkzeug==2.3.7
zipp==3.20.2
